PPT/requirements.txt
python-pptx
openai
pydantic
python-dotenv

